chore(structure): structure app
- create folders

[Finishes #169195228]
